refactor: organize `link.ts` into module
