+++ scripts/link/noop.ts
@@ -1,0 +1,2 @@
+/** does nothing. */
+export const noop = () => {}
